+++ src/core/databasemanager.h
@@ -19,6 +19,7 @@ public:
     void insertDownload(const downloadInformations &info);
     void downloadFinished(const QString &id, const QString &status);
     void removeDownload(const QString &id);
+    QVector<downloadInformations> getDownloads();
 
 private:
     // Functions